In StorageImpl Class created 2 arrays, constant int for storage length and arraySize.
created the constructor for StorageImpl.
implemented put() method, where with help of for loop I've managed to check if key == null and if there is already a key value in the keyArr.
implemented get() method, with the for loop where and the same logic as put method.
implemented size() method, with for loop to check if elements in the array are not equal to null and return the size of the array.
////////////////////////////////////////////////////////////
Changes after review were done.