Added in the StorageImpl class: - ENTRY_ITEMS_NUMBER constant, which limits the number of elements in the entries array; - An array of Entry<K, V>[] entries where objects will be stored; - An int size variable that will store the actual number of objects in the entries array.
Added StorageImpl() method to initialize entries and size.

The put() method from the Storage interface is overridden, which takes the values K key and V value.
Inside, a loop is declared that iterates through the entries array in search of duplicate entries.
If the key from entries matches the incoming one (a separate isEqualKeys() method has been made for this), then we overwrite the value and interrupt the loop with the return.
If the condition inside the loop did not work, we write the values to the entries array at the index (value of size) and increment the size variable.
Adding elements to the entries is wrapped in a try-catch block in case the user tries to add more entries than the entries array can hold.

Override get() method from the Storage interface, which takes a K key value.
Looping through the entries array, if the key values match (isEqualKeys()), return the corresponding value.
Otherwise, we return null.

Override the size() method from the Storage interface, which returns the value of size.

Added the boolean isEqualKeys() method, which accepts a key from the implementation and a key from an element of the entries array. It works according to the following principle:
1. If the incoming key and the key from entries are both null - true;
2. If only one of the keys is null - false;
3. If the keys are not null, but their values are the same - true, otherwise false.
For this implementation, the loop must be fori so that you can limit it to the actual number of entries using the size variable. Otherwise, isEqualKeys() will not work correctly due to the default initialization of empty arrays.