Created two arrays and size variable : First array "keys" store keys value; Second array "values" store value; I use "size" variable to check how many keys store in array
Override equals method to check does the key exist in the keys array.If the key has been found, I replace its value in the values array.

If the key is not present, I add a new key and its corresponding value to the end of the keys and values arrays, using the variable size to determine the position in the array.

In the "get()" method, I locate the key in the keys array in the same way as in the "put()" method, and return the corresponding value from the values array for the found key.

In the size method, I return size.