In this implementation, we use two arrays: keys to store the keys and values to store the corresponding values. The size variable keeps track of the number of elements in the storage.
When you add a new key-value pair using the put method, it checks if the key already exists in the storage. If it does, the existing value is replaced. If not, it checks if the storage is already full before adding the new key-value pair.

The get method retrieves the value associated with the given key if it exists, otherwise, it returns null.

The size method simply returns the current size of the storage.

Note: Since Java does not allow the creation of generic arrays directly, we use type casting with Object arrays and suppress the unchecked warning.